refactor(ChBottomSheet): Rename component
Fix mistypo in component's name
